Move WindowedKeyValueCache to separate files.
