+++ src/models/kv_cache.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "model.h"
 #include "static_buffer.h"
 
 namespace Generators {
@@ -100,39 +101,7 @@ struct CrossCache {
   std::vector<std::string> input_name_strings_, output_name_strings_;
 };
 
-struct WindowedKeyValueCache : KeyValueCache {
-  WindowedKeyValueCache(State& state);
-
-  void Add() override;
-  void AddEncoder() override {
-    throw std::runtime_error("WindowedKeyValueCache does not support AddEncoder.");
-  };
-  void Update(DeviceSpan<int32_t> beam_indices, int current_length) override;
-  void RewindTo(size_t index) override {
-    throw std::runtime_error("WindowedKeyValueCache does not support RewindTo.");
-  }
-
- private:
-  void Slide();
-
-  State& state_;
-  const Model& model_{state_.model_};
-  int layer_count_{};
-  int window_size_{};
-  size_t num_windows_{};
-  size_t window_index_{};
-  size_t input_index_{~0U}, output_index_{~0U};
-
-  std::array<int64_t, 4> key_cache_shape_in_, key_cache_shape_out_;
-  std::array<int64_t, 4> value_cache_shape_in_, value_cache_shape_out_;
-  ONNXTensorElementDataType type_;
-
-  std::vector<std::unique_ptr<OrtValue>> key_caches_in_, value_caches_in_;
-  std::vector<std::unique_ptr<OrtValue>> key_caches_out_, value_caches_out_;
-  std::vector<std::string> input_name_strings_, output_name_strings_;
-
-  bool is_first_update_{true};
-};
+std::string ComposeKeyValueName(const std::string& template_string, int index);
 
 std::unique_ptr<KeyValueCache> CreateKeyValueCache(State& state);
